# Optimized event_based_risk