--- a/openquake/risklib/scientific.py
+++ b/openquake/risklib/scientific.py
@@ -32,8 +32,10 @@ from scipy import interpolate, stats
 F64 = numpy.float64
 F32 = numpy.float32
 U32 = numpy.uint32
+U64 = numpy.uint64
 U16 = numpy.uint16
 U8 = numpy.uint8
+TWO32 = 2 ** 32
 KNOWN_CONSEQUENCES = ['loss', 'ins_loss', 'losses', 'collapsed', 'injured',
                       'fatalities', 'homeless']
 
@@ -1482,6 +1484,30 @@ def get_agg_value(consequence, agg_values, agg_id, loss_type):
         return aval[loss_type]
     else:
         raise NotImplementedError(consequence)
+
+
+# ########################### u64_to_eal ################################# #
+
+def u64_to_eal(u64):
+    """
+    Convert an unit64 into a triple (eid, aid, lid)
+
+    >>> u64_to_eal(42949673216001)
+    (10000, 1000, 1)
+    """
+    eid, x = divmod(u64, TWO32)
+    aid, lid = divmod(x, 256)
+    return eid, aid, lid
+
+
+def eal_to_u64(eid, aid, lid):
+    """
+    Convert a triple (eid, aid, lid) into an uint64:
+
+    >>> eal_to_u64(10000, 1000, 1)
+    42949673216001
+    """
+    return U64(eid * TWO32) + U64(aid * 256) + U64(lid)
 
 
 if __name__ == '__main__':