--- a/openquake/risklib/scientific.py
+++ b/openquake/risklib/scientific.py
@@ -1407,13 +1407,13 @@ def insurance_losses(asset_df, losses_by_rl, policy_df):
         if len(policy):
             new = out.copy()
             new['variance'] = 0.
-            for (eid, aid), df in out.iterrows():
+            for (eid, aid), ser in out.iterrows():
                 asset = asset_df.loc[aid]  # aid==ordinal
                 avalue = asset['value-' + lt]
                 policy_idx = asset['policy']
                 ded = policy.loc[policy_idx].deductible
                 lim = policy.loc[policy_idx].insurance_limit
-                ins = insured_losses(df.loss, ded * avalue, lim * avalue)
+                ins = insured_losses(ser.loss, ded * avalue, lim * avalue)
                 new.loc[eid, aid]['loss'] = ins
             losses_by_rl[riskid, lt + '_ins'] = new