By removing .set_index and .reset_index we get some speedup. For instance, for the slovenia calculation in oq-risk-tests:
```
# before
| calc_9419, maxmem=2.7 GB     | time_sec | memory_mb | counts |
|------------------------------+----------+-----------+--------|
| total event_based_risk       | 33.3     | 33.4      | 16     |
| computing risk               | 19.0     | 0.0       | 2_096  |
| reading data                 | 3.32491  | 19.3      | 16     |
| averaging losses             | 2.79689  | 0.0       | 3_155  |
| aggregating losses           | 2.54641  | 0.0       | 3_155  |
# after
| calc_9378, maxmem=2.7 GB     | time_sec | memory_mb | counts |
|------------------------------+----------+-----------+--------|
| total event_based_risk       | 26.7     | 33.1      | 16     |
| computing risk               | 14.8     | 0.0       | 2_096  |
| reading data                 | 3.25910  | 19.3      | 16     |
| averaging losses             | 2.80203  | 0.0       | 3_155  |
| aggregating losses           | 2.45717  | 0.0       | 3_155  |
```
Also improved a bit the calculation of the insurance losses and added two helper functions that might be useful in the future. Part of #7886